Heavily refactored buyer integration test to ensure paypal purchases and ones with IPN go through smoothly.
Ensured also that finished orders never can have their items modified through order.empty! which was casuing issue 220.

This closes issue 219 and issue 220.

+++ plugins/substruct/app/models/order.rb
@@ -520,8 +520,9 @@ class Order < ActiveRecord::Base
     self.order_line_items.size == 0
   end
 
-  # Removes all items from order
+  # Removes all items from order unless it's been placed.
   def empty!
+    return if self.is_complete?
     self.order_line_items.destroy_all
   end
